COMP: Fix clang v. 5 compilation on Linux
Hopefully helps for issue #31

+++ k3dsdk/gil/boost/uuid/seed_rng.hpp
@@ -140,7 +140,7 @@ private:
         }
 
         {
-            unsigned int rn[] = { std::rand(), std::rand(), std::rand() };
+            unsigned int rn[] = { static_cast<unsigned int>(std::rand()), static_cast<unsigned int>(std::rand()), static_cast<unsigned int>(std::rand()) };
             sha.process_bytes( (unsigned char const*)rn, sizeof( rn ) );
         }
